+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 Change Log
 ==========
+0.0.5
+-----
+  * Added dependency existence check to adp:package:install
+  * Fixed bug in adp:source:deploy
+  * Fixed help on tooling:query command
+
 0.0.4
 -----
   * Enhancements to package install command